Allow for selector expressions in URLs.
This enables a similar behavior to `jQuery.load()', wherein a selector expression can be added to the URL, causing only content matching that expression to be used as dialog content. For example:

    http://www.example.com/ #content

The entire document at "http://www.example.com" will be retrieved, but only the content matching the selector "#content" will be used as dialog content.

This also exposes the property "selector" (in addition to the already-existing "href" property).

See the documentation for `jQuery.load()' for a more detailed description about this behavior:

    http://api.jquery.com/load/