Enforce redraw of fancybox wrap
In Webkit-based browsers, if your mousecursor is exactly at the position
of the next-button while fancybox slides in, you have to click twice to
get to the next image because the first click gets eaten by Webkit to
check the :hover state of the 'a' element used for the next-button.

If we enforce a redraw by doing a hide().show(0), Webkit will reevaluate
the :hover state and the first click will trigger an click event like
expected.