Add support for Turbolinks
Turbolinks (a core part of the well-know Rails framework) is not compatible with fancybox: it's ok when we load the page for the first time, but brokens when we change between pages (by clicking on links).
The problem is because part of fancybox initialization process is executed when the "ready" event is triggered, but Turbolinks doesn't trigger "ready" when we change between pages - so some initialization is never executed. Instead trigger "ready", Turbolinks triggers a custom event called "page:load".
This PR listen to this event too and executes the initialization process correctly, without affect another frameworks/existent/legacy code.